Removed unnecessary debug logs
Cleaned up.
