--- Penguin Tower Defense/Assets/Scrips/Lives/BaseEnemy.cs
@@ -1,10 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class BaseEnemy : MonoBehaviour
-{
-    protected int damage;
-    protected int health;
-    protected float speed;
-}
