--- Penguin Tower Defense/Assets/Scrips/Lives/BaseEnemy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: df71632dc25f4b24e893edd31ee1c62e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
