--- Penguin Tower Defense/Assets/Scrips/Lives/GameOver.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class GameOver : MonoBehaviour
-{
-    public void EndGame()
-    {
-        Debug.Log("Game Over!");
-    }
-
-}
