+++ Penguin Tower Defense/Assets/Scripts/Lives/GameOver.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class GameOver : MonoBehaviour
+{
+    public void EndGame()                               //Ends the game (not finished)
+    {
+        Debug.Log("Game Over!");
+    }
+
+}
